+++ clang-tools-extra/docs/ReleaseNotes.rst
@@ -166,6 +166,11 @@ New checks
 - New :doc:`bugprone-assignment-in-if-condition
   <clang-tidy/checks/bugprone/assignment-in-if-condition>` check.
 
+- New :doc:`misc-structured-namespace
+  <clang-tidy/checks/misc/structured-namespace>` check.
+
+  FIXME: add release notes.
+
 - New :doc:`modernize-use-ranges
   <clang-tidy/checks/modernize/use-ranges>` check.